Move staff members to external
