+++ node/src/main/kotlin/net/corda/node/services/statemachine/hospital/external/DeadlockNurse.kt
@@ -1,4 +1,4 @@
-package net.corda.node.services.statemachine.hospital
+package net.corda.node.services.statemachine.hospital.external
 
 import net.corda.node.services.statemachine.Chronic
 import net.corda.node.services.statemachine.Diagnosis
@@ -8,6 +8,8 @@ import net.corda.node.services.statemachine.Staff
 import net.corda.node.services.statemachine.StateMachineState
 import net.corda.node.services.statemachine.mentionsThrowable
 import java.sql.SQLException
+
+// [DeadlockNurse] could go in corda transactions module?
 
 /**
  * SQL Deadlock detection.
