+++ node/src/main/kotlin/net/corda/node/services/statemachine/hospital/external/DuplicateInsertSpecialist.kt
@@ -1,4 +1,4 @@
-package net.corda.node.services.statemachine.hospital
+package net.corda.node.services.statemachine.hospital.external
 
 import net.corda.node.services.statemachine.Diagnosis
 import net.corda.node.services.statemachine.FlowFiber
@@ -7,6 +7,8 @@ import net.corda.node.services.statemachine.Staff
 import net.corda.node.services.statemachine.StateMachineState
 import net.corda.node.services.statemachine.mentionsThrowable
 import org.hibernate.exception.ConstraintViolationException
+
+// [DuplicateInsertSpecialist] could go in corda transactions module?
 
 /**
  * Primary key violation detection for duplicate inserts.  Will detect other constraint violations too.
